shell scripts refactored

+++ scripts/clean_environment.sh
@@ -8,16 +8,23 @@
 # CALL FROM COMMAND LINE: . ./clean_environment.sh && clean_environment                      #
 ##############################################################################################
 
-clean_environment()
-{
-    printf "${YELLOW}Cleaning developing environment at '${PWD}'...${NORMAL}\n"
-    find . -type f -name "*.py[co]" -delete
-    find . -type d -name "__pycache__" -delete
-    find . -type f -name ".coverage" -delete
-    rm -rf ands.egg-info
-    rm -rf venv
-    printf "${GREEN}Done.${NORMAL}\n"
-}
+export ALREADY_SOURCED_CLEAN_ENVIRONMENT
 
-. ./_source_script.sh
-_source_script scripts colors
+if [ -z "${ALREADY_SOURCED_CLEAN_ENVIRONMENT}" ]
+then
+    clean_environment()
+    {
+        printf "%sCleaning developing environment at '%s'...%s" "${YELLOW}" "${PWD}" "${NORMAL}"
+        find . -type f -name "*.py[co]" -delete
+        find . -type d -name "__pycache__" -delete
+        find . -type f -name ".coverage" -delete
+        rm -rf ands.egg-info
+        rm -rf venv
+        printf "%s done.%s\n" "${GREEN}" "${NORMAL}"
+    }
+
+    . ./_source_script.sh
+    _source_script scripts colors
+
+    ALREADY_SOURCED_CLEAN_ENVIRONMENT="true"
+fi
